+++ lib/vagrant/hosts/bsd.rb
@@ -4,10 +4,15 @@ module Vagrant
     class BSD < Base
       include Util
       include Util::Retryable
+      include Util::Sh
 
       def nfs?
         retryable(:tries => 10, :on => TypeError) do
-          system("which nfsd > /dev/null 2>&1")
+          _, status = sh("which nfsd")
+
+          # Sometimes the status is nil for some reason. In that case, force a retry
+          raise TypeError.new("Bad status code") if !status
+          status.success?
         end
       end
 
@@ -25,22 +30,22 @@ module Vagrant
         output.split("\n").each do |line|
           # This should only ask for administrative permission once, even
           # though its executed in multiple subshells.
-          system(%Q[sudo su root -c "echo '#{line}' >> /etc/exports"])
+          sh(%Q[sudo su root -c "echo '#{line}' >> /etc/exports"])
         end
 
         # We run restart here instead of "update" just in case nfsd
         # is not starting
-        system("sudo nfsd restart")
+        sh("sudo nfsd restart")
       end
 
       def nfs_cleanup
         return if !File.exist?("/etc/exports")
-        system("cat /etc/exports | grep 'VAGRANT-BEGIN: #{env.vm.uuid}' > /dev/null 2>&1")
+        _, status = sh("cat /etc/exports | grep 'VAGRANT-BEGIN: #{env.vm.uuid}'")
 
-        if $?.to_i == 0
+        if status.success?
           # Use sed to just strip out the block of code which was inserted
           # by Vagrant
-          system("sudo sed -e '/^# VAGRANT-BEGIN: #{env.vm.uuid}/,/^# VAGRANT-END: #{env.vm.uuid}/ d' -i bak /etc/exports")
+          sh("sudo sed -e '/^# VAGRANT-BEGIN: #{env.vm.uuid}/,/^# VAGRANT-END: #{env.vm.uuid}/ d' -i bak /etc/exports")
         end
       end
     end