+++ lib/vagrant/util.rb
@@ -7,6 +7,7 @@ module Vagrant
     autoload :Platform,                  'vagrant/util/platform'
     autoload :ResourceLogger,            'vagrant/util/resource_logger'
     autoload :Retryable,                 'vagrant/util/retryable'
+    autoload :Sh,                        'vagrant/util/sh'
     autoload :StackedProcRunner,         'vagrant/util/stacked_proc_runner'
     autoload :TemplateRenderer,          'vagrant/util/template_renderer'
   end